create english user
